+++ CHANGELOG.md
@@ -8,7 +8,8 @@ app.
 
 ## [Unreleased]
 
-(Nothing yet)
+- Fixed a bug where the popup would sometimes pin when mouse interaction was
+  disabled and hold-to-show keys were configured.
 
 ## [1.12.2] - 2022-09-10